sincronitzar archetype per #79

+++ archetype/src/main/resources/archetype-resources/__rootArtifactId__-ejb/src/main/java/ejb/facade/UnitatOrganicaServiceFacadeBean.java
@@ -5,15 +5,17 @@ package ${package}.ejb.facade;
 
 import ${package}.commons.utils.Constants;
 import ${package}.ejb.converter.UnitatOrganicaConverter;
-import ${package}.ejb.interceptor.ExceptionInterceptor;
+import ${package}.ejb.interceptor.ExceptionTranslate;
 import ${package}.ejb.interceptor.Logged;
+import ${package}.ejb.repository.ProcedimentRepository;
 import ${package}.ejb.repository.UnitatOrganicaRepository;
 import ${package}.persistence.model.UnitatOrganica;
 import ${package}.service.exception.RecursNoTrobatException;
-import ${package}.service.exception.UnitatOrganicaDuplicadaExeption;
+import ${package}.service.exception.UnitatDuplicadaException;
+import ${package}.service.exception.UnitatTeProcedimentsException;
 import ${package}.service.facade.UnitatOrganicaServiceFacade;
 import ${package}.service.model.Ordre;
-import ${package}.service.model.Page;
+import ${package}.service.model.Pagina;
 import ${package}.service.model.UnitatOrganicaDTO;
 
 import javax.annotation.security.RolesAllowed;
@@ -22,9 +24,6 @@ import javax.ejb.Stateless;
 import javax.ejb.TransactionAttribute;
 import javax.ejb.TransactionAttributeType;
 import javax.inject.Inject;
-import javax.interceptor.Interceptors;
-import javax.validation.executable.ValidateOnExecution;
-
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -32,28 +31,30 @@ import java.util.Optional;
 /**
  * Implementació dels casos d'ús de manteniment de Unitats Orgàniques.
  * És responsabilitat d'aquesta capa definir el limit de les transaccions i la seguretat.
- *
+ * Les excepcions específiques es llancen mitjançant l'{@link ExceptionTranslate} que transforma
+ * els errors JPA amb les excepcions de servei com la {@link RecursNoTrobatException}
  * @author areus
  */
 @Logged
-@Stateless
-@Local(UnitatOrganicaServiceFacade.class)
+@ExceptionTranslate
+@Stateless @Local(UnitatOrganicaServiceFacade.class)
 @TransactionAttribute(TransactionAttributeType.REQUIRES_NEW)
 @RolesAllowed(Constants.${prefixuppercase}_ADMIN)
-@ValidateOnExecution
-@Interceptors(ExceptionInterceptor.class)
 public class UnitatOrganicaServiceFacadeBean implements UnitatOrganicaServiceFacade {
 
     @Inject
     private UnitatOrganicaRepository repository;
 
     @Inject
+    private ProcedimentRepository procedimentRepository;
+
+    @Inject
     private UnitatOrganicaConverter converter;
 
     @Override
-    public Long create(UnitatOrganicaDTO dto) throws UnitatOrganicaDuplicadaExeption {
+    public Long create(UnitatOrganicaDTO dto) throws UnitatDuplicadaException {
         if (repository.findByCodiDir3(dto.getCodiDir3()).isPresent()) {
-            throw new UnitatOrganicaDuplicadaExeption(dto.getCodiDir3());
+            throw new UnitatDuplicadaException(dto.getCodiDir3());
         }
 
         UnitatOrganica unitat = converter.toEntity(dto);
@@ -62,36 +63,34 @@ public class UnitatOrganicaServiceFacadeBean implements UnitatOrganicaServiceFac
     }
 
     @Override
-    public void update(UnitatOrganicaDTO dto) throws UnitatOrganicaDuplicadaExeption {
-        Optional<UnitatOrganica> opUnitat = repository.findByCodiDir3(dto.getCodiDir3());
-        if (opUnitat.isPresent() && !opUnitat.get().getId().equals(dto.getId())) {
-            throw new UnitatOrganicaDuplicadaExeption(dto.getCodiDir3());
-        }
-
-        UnitatOrganica unitat = opUnitat.orElse(repository.getReference(dto.getId()));
+    public void update(UnitatOrganicaDTO dto) throws RecursNoTrobatException {
+        UnitatOrganica unitat = repository.getReference(dto.getId());
         converter.updateFromDTO(unitat, dto);
-        repository.update(unitat);
     }
 
     @Override
-    public void delete(Long id) {
+    public void delete(Long id) throws UnitatTeProcedimentsException, RecursNoTrobatException {
+        if (procedimentRepository.countByUnitat(id) > 0L) {
+            throw new UnitatTeProcedimentsException(id);
+        }
         UnitatOrganica unitat = repository.getReference(id);
         repository.delete(unitat);
     }
 
     @Override
-    public UnitatOrganicaDTO findById(Long id) throws RecursNoTrobatException {
-        UnitatOrganica unitat = repository.findById(id).orElseThrow(RecursNoTrobatException::new);
-        return converter.toDTO(unitat);
+    public Optional<UnitatOrganicaDTO> findById(Long id) {
+        UnitatOrganica unitat = repository.findById(id);
+        UnitatOrganicaDTO unitatOrganicaDTO = converter.toDTO(unitat);
+        return Optional.ofNullable(unitatOrganicaDTO);
     }
 
     @Override
-    public Page<UnitatOrganicaDTO> findFiltered(int firstResult, int maxResult, Map<String, Object> filters,
-                                                List<Ordre> ordenacio) {
+    public Pagina<UnitatOrganicaDTO> findFiltered(int firstResult, int maxResult, Map<String, Object> filters,
+                                                  List<Ordre> ordenacio) {
 
         List<UnitatOrganicaDTO> items = repository.findPagedByFilterAndOrder(firstResult, maxResult, filters, ordenacio);
         long total = repository.countByFilter(filters);
 
-        return new Page<>(items, total);
+        return new Pagina<>(items, total);
     }
 }
